# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "mts-server",
    "description": "MTS - An AI agent by Magezi Tech Solutions",
    "contact": {
      "name": "Magezi Tech Solutions"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.16.0"
  },
  "paths": {
...
    "/search/filenames": {
      "post": {
        "tags": [
          "super::routes::search"
        ],
        "operationId": "search_filenames",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchFilenamesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Filename search completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchFilenamesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/search/files": {
      "post": {
        "tags": [
          "super::routes::search"
        ],
        "operationId": "search_files",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchFilesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchFilesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/search/replace": {
      "post": {
        "tags": [
          "super::routes::search"
        ],
        "operationId": "replace_in_files",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replace operation completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid regex pattern"
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FilenameMatch": {
        "type": "object",
        "required": [
          "path",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "ReplaceRequest": {
        "type": "object",
        "required": [
          "query",
          "replacement",
          "workingDir"
        ],
        "properties": {
          "caseSensitive": {
            "type": "boolean"
          },
          "excludePattern": {
            "type": "string",
            "nullable": true
          },
          "filePaths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "includePattern": {
            "type": "string",
            "nullable": true
          },
          "query": {
            "type": "string"
          },
          "replacement": {
            "type": "string"
          },
          "useRegex": {
            "type": "boolean"
          },
          "wholeWord": {
            "type": "boolean"
          },
          "workingDir": {
            "type": "string"
          }
        }
      },
      "ReplaceResponse": {
        "type": "object",
        "required": [
          "filesModified",
          "totalReplacements"
        ],
        "properties": {
          "filesModified": {
            "type": "integer",
            "minimum": 0
          },
          "totalReplacements": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "SearchFilenamesRequest": {
        "type": "object",
        "required": [
          "query",
          "workingDir"
        ],
        "properties": {
          "maxResults": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "query": {
            "type": "string"
          },
          "workingDir": {
            "type": "string"
          }
        }
      },
      "SearchFilenamesResponse": {
        "type": "object",
        "required": [
          "matches",
          "truncated"
        ],
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilenameMatch"
            }
          },
          "truncated": {
            "type": "boolean"
          }
        }
      },
      "SearchFilesRequest": {
        "type": "object",
        "required": [
          "query",
          "workingDir"
        ],
        "properties": {
          "caseSensitive": {
            "type": "boolean"
          },
          "contextLines": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "excludePattern": {
            "type": "string",
            "nullable": true
          },
          "includePattern": {
            "type": "string",
            "nullable": true
          },
          "maxResults": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "query": {
            "type": "string"
          },
          "useRegex": {
            "type": "boolean"
          },
          "wholeWord": {
            "type": "boolean"
          },
          "workingDir": {
            "type": "string"
          }
        }
      },
      "SearchFilesResponse": {
        "type": "object",
        "required": [
          "matches",
          "totalFiles",
          "totalMatches",
          "truncated"
        ],
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchMatch"
            }
          },
          "totalFiles": {
            "type": "integer",
            "minimum": 0
          },
          "totalMatches": {
            "type": "integer",
            "minimum": 0
          },
          "truncated": {
            "type": "boolean"
          }
        }
      },
      "SearchMatch": {
        "type": "object",
        "required": [
          "filePath",
          "lineNumber",
          "column",
          "lineText",
          "contextBefore",
          "contextAfter"
        ],
        "properties": {
          "column": {
            "type": "integer",
            "minimum": 0
          },
          "contextAfter": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contextBefore": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filePath": {
            "type": "string"
          },
          "lineNumber": {
            "type": "integer",
            "minimum": 0
          },
          "lineText": {
            "type": "string"
          }
        }
      },
...
    }
  }
}